Refactor: [FE] 주석 및 파일 구조 정리
