--- client/src/pages/Main/subpages/Expenses/Expense.mockList.ts
@@ -1,24 +1,0 @@
-//TODO: mock list 제거
-export const mockList = [
-  {
-    businessName: "(주)배달의민족",
-    amounts: 20000,
-    date: "2024-01-16 17:50:00",
-    category: "식비",
-    owner: "만두",
-  },
-  {
-    businessName: "(주)굿모닝마트 교대역점",
-    amounts: 30000,
-    date: "2024-01-16 17:50:00",
-    category: "식비",
-    owner: "만두",
-  },
-  {
-    businessName: "(주)영풍문고코엑스점",
-    amounts: 16500,
-    date: "2024-01-16 17:50:00",
-    category: "문화생활",
-    owner: "만두",
-  },
-];
